refactor(slack-bot): extract config.js from index.js (behavior-preserving)
